requirements.txt
streamlit==1.37.1
feedparser==6.0.11
pandas==2.2.2
pyyaml==6.0.2
python-dateutil==2.9.0.post0
beautifulsoup4==4.12.3
lxml==5.2.2
